[최준] discount(추가,수정,삭제,리스트 기능완성)css작업필요, product(productList)css수정, main(menuBar, empMain, home)수정,

+++ src/main/java/vo/Discount.java
@@ -35,7 +35,7 @@ public class Discount {
 	public double getDiscountRate() {
 		return discountRate;
 	}
-	public void setDiscountRate(int discountRate) {
+	public void setDiscountRate(Double discountRate) {
 		this.discountRate = discountRate;
 	}
 	public String getCreatedate() {
